requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.24.1,<0.25
uwsgi>=2.0.20<2.1
